Package it into a single file
封装成一个单独文件，使用更方便。调整了左边栏与主界面没对齐的问题
