+++ QQConfiguration.xcodeproj/xcuserdata/lulinxuan.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -1,0 +1,27 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>SchemeUserState</key>
+	<dict>
+		<key>QQConfiguration.xcscheme</key>
+		<dict>
+			<key>orderHint</key>
+			<integer>0</integer>
+		</dict>
+	</dict>
+	<key>SuppressBuildableAutocreation</key>
+	<dict>
+		<key>CE19AB6B1ACE2CE300A221EC</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>CE19AB801ACE2CE300A221EC</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+	</dict>
+</dict>
+</plist>
